refactor(datasource): 将活动ID类型从Int改为Long并完善图片管理功能
- 将Activities实体类中的id字段从Int改为Long类型
- 更新ActivitiesRepository和ActivitiesService接口中的泛型参数为Long
- 修改所有相关服务实现类以支持Long类型的活动ID
- 在控制器层将IntIdRequest和IntIdsRequest替换为LongIdRequest和LongIdsRequest
- 更新请求对象中activityId字段的类型为Long
- 修改ControllerSupport中的requireActivityId方法参数类型为Long
- 完善图片管理HTML界面，添加活动ID输入验证
- 扩展ImageService接口增加图片上传、软删除等功能
- 实现完整的图片上传、存储、访问功能，包括SHA-256校验和文件类型验证
- 配置图片存储路径规范化和安全性检查
- 更新application.yaml中图片配置的命名规范
- 修改所有依赖活动ID的数据访问方法以支持Long类型参数

+++ src/main/kotlin/top/foxball/receptionbackendsystem/controller/ControllerSupport.kt
@@ -57,5 +57,5 @@ abstract class ControllerSupport(
         )
     }
 
-    protected fun requireActivityId(activityId: Int?): Int? = activityId
+    protected fun requireActivityId(activityId: Long?): Long? = activityId
 }
