+++ src/main/kotlin/top/foxball/receptionbackendsystem/datasource/jdbc/PersonRepository.kt
@@ -7,7 +7,7 @@ interface PersonRepository : ReceptionRepository<Person, Int> {
     /**
      * 查询指定活动下的全部人员。
      */
-    fun findByActivityId(activityId: Int): List<Person>
+    fun findByActivityId(activityId: Long): List<Person>
 
     /**
      * 根据人员姓名模糊查询。
@@ -17,5 +17,5 @@ interface PersonRepository : ReceptionRepository<Person, Int> {
     /**
      * 查询指定活动下指定单位的人员。
      */
-    fun findByActivityIdAndUnit(activityId: Int, unit: String): List<Person>
+    fun findByActivityIdAndUnit(activityId: Long, unit: String): List<Person>
 }
