+++ src/main/kotlin/top/foxball/receptionbackendsystem/datasource/jdbc/PromptServiceRepository.kt
@@ -7,5 +7,5 @@ interface PromptServiceRepository : ReceptionRepository<PromptService, Int> {
     /**
      * 查询指定活动下的全部提示服务配置。
      */
-    fun findByActivityId(activityId: Int): List<PromptService>
+    fun findByActivityId(activityId: Long): List<PromptService>
 }
