+++ src/main/kotlin/top/foxball/receptionbackendsystem/service/ActivityEntityService.kt
@@ -1,5 +1,5 @@
 package top.foxball.receptionbackendsystem.service
 
 interface ActivityEntityService<T : Any, ID : Any> : ReceptionService<T, ID> {
-    fun findByActivityId(activityId: Int): List<T>
+    fun findByActivityId(activityId: Long): List<T>
 }
