+++ src/main/kotlin/top/foxball/receptionbackendsystem/service/CarService.kt
@@ -3,5 +3,5 @@ package top.foxball.receptionbackendsystem.service
 import top.foxball.receptionbackendsystem.datasource.jdbc.Car
 
 interface CarService : ActivityEntityService<Car, Int> {
-    fun findByActivityIdAndCarNumber(activityId: Int, carNumber: Long): Car?
+    fun findByActivityIdAndCarNumber(activityId: Long, carNumber: Long): Car?
 }
